requirements.txt
pathlib~=1.0.1
python-dotenv~=1.0.1
bs4~=0.0.2
beautifulsoup4~=4.13.3
docling~=2.28.0
markitdown~=0.1.1
pdfkit~=1.0.0
requests~=2.32.3
Markdown~=3.7
langchain~=0.3.21
psycopg2-binary~=2.9.10
langchain-google-genai~=2.1.1
langchain-community~=0.3.20
fitz~=0.0.1.dev2
pytesseract~=0.3.13
pdf2image~=1.17.0
tqdm~=4.67.1
Scrapy~=2.12.0
jmespath~=1.0.1
pandas~=2.2.3
numpy~=2.2.4
pyxnat~=1.6.3
defusedxml~=0.7.1
ipython~=8.34.0
pillow~=11.1.0
olefile~=0.47
pytest~=8.3.5
lxml~=5.3.1
soupsieve~=2.6
packaging~=24.2
typing_extensions~=4.12.2
html5lib~=1.1
bleach~=6.2.0
h11~=0.14.0
pip~=23.2.1
emoji~=2.14.1
configparser~=7.2.0
attrs~=25.3.0
wheel~=0.41.2
protobuf~=5.29.4
pyOpenSSL~=25.0.0
cryptography~=44.0.2
tornado~=6.4.2
Jinja2~=3.1.6
rsa~=4.9
pyasn1~=0.6.1
pyzmq~=26.3.0
pexpect~=4.9.0
cffi~=1.17.1
python-bidi~=0.6.6
setuptools~=68.2.0
comm~=0.2.2
traitlets~=5.14.3
dill~=0.3.9
python-docx~=1.1.2
networkx~=3.4.2
nipype~=1.10.0
PyMuPDF~=1.25.4
fqdn~=1.5.1
grpcio~=1.71.0
parso~=0.8.4
jedi~=0.19.2
cssselect~=1.3.0
nltk~=3.9.1
regex~=2024.11.6
scipy~=1.15.2
click~=8.1.8
pyparsing~=3.2.3
python-pptx~=1.0.2
XlsxWriter~=3.2.2
prov~=2.0.1
rdflib~=6.3.2
pydot~=3.0.4
python-dateutil~=2.9.0.post0
pytz~=2025.2
rich~=13.9.4
Pygments~=2.19.1
markdown-it-py~=3.0.0
notebook~=7.3.3
PyYAML~=6.0.2
idna~=3.10
multidict~=6.2.0
propcache~=0.3.0
importlib_resources~=6.5.2
exceptiongroup~=1.2.2
sniffio~=1.3.1
arrow~=1.3.0
babel~=2.17.0
ffmpy~=0.5.0
httpcore~=1.0.7
httpx~=0.28.1
zstandard~=0.23.0
json5~=0.10.0
marko~=2.1.2
mdurl~=0.1.2
mpire~=2.10.2
MarkupSafe~=3.0.2
multiprocess~=0.70.17
ninja~=1.11.1.4
charset-normalizer~=3.4.1
overrides~=7.7.0
proto-plus~=1.26.1
pypdf~=5.4.0
torch~=2.6.0
mpmath~=1.3.0
sympy~=1.13.1
platformdirs~=4.3.7
decorator~=5.2.1
tabulate~=0.9.0
onnxruntime~=1.21.0
triton~=3.2.0
torchvision~=0.21.0
psutil~=7.0.0
filelock~=3.18.0
urllib3~=2.3.0
fsspec~=2025.3.0
typer~=0.12.5
shellingham~=1.5.4
w3lib~=2.3.1
argon2-cffi-bindings~=21.2.0
webencodings~=0.5.1
tinycss2~=1.4.0
aiohttp~=3.11.14
yarl~=1.18.3
pyasn1_modules~=0.4.1
cachetools~=5.5.2
certifi~=2025.1.31
grpcio-status~=1.71.0
gradio_client~=1.8.0
gradio~=5.23.1
tomlkit~=0.13.2
semantic-version~=2.10.0
huggingface-hub~=0.29.3
fastapi~=0.115.12
starlette~=0.46.1
anyio~=4.9.0
orjson~=3.10.16
groovy~=0.1.2
pydub~=0.25.1
pydantic~=2.10.6
transformers~=4.50.0
uvicorn~=0.34.0
pydantic_core~=2.27.2
aiofiles~=24.1.0
safehttpx~=0.1.6
joblib~=1.4.2
magika~=0.6.1
nibabel~=5.3.2
looseversion~=1.3.0
simplejson~=3.20.1
traits~=7.0.2
acres~=0.3.0
puremagic~=1.28
openpyxl~=3.1.5
SQLAlchemy~=2.0.39
parsel~=1.10.0
isodate~=0.6.1
Twisted~=24.11.0
itemadapter~=0.11.0
queuelib~=1.7.0
PyDispatcher~=2.0.7
itemloaders~=1.3.2
tldextract~=5.1.3
stack-data~=0.6.3
matplotlib-inline~=0.1.7
python-magic~=0.4.27
jupyter_client~=8.6.3
prompt_toolkit~=3.0.50
wcwidth~=0.2.13
pluggy~=1.5.0
iniconfig~=2.1.0
async-timeout~=4.0.3
aiosignal~=1.3.2
frozenlist~=1.5.0
python-multipart~=0.0.20
aiohappyeyeballs~=2.6.1
Automat~=24.8.1
backoff~=2.2.1
chardet~=5.2.0
ci-info~=0.3.0
debugpy~=1.8.13
greenlet~=3.1.1
rtree~=1.4.0
pylatexenc~=2.10
pypdfium2~=4.30.1
filetype~=1.2.0
pydantic-settings~=2.8.1
pyclipper~=1.3.0.post6
shapely~=2.0.7
easyocr~=1.7.2
six~=1.17.0
scikit-image~=0.25.2
websockets~=15.0.1
imageio~=2.37.0
tifffile~=2025.3.13
ptyprocess~=0.7.0
psycopg~=3.2.6
psycopg-pool~=3.2.6
lazy_loader~=0.4
incremental~=24.7.2
constantly~=23.10.4
hyperlink~=21.0.0
httplib2~=0.22.0
nbformat~=5.10.4
jupyter_core~=5.7.2
nbclient~=0.10.2
fastjsonschema~=2.21.1
jsonschema~=4.23.0
jupyter_server~=2.15.0
jupyterlab~=4.3.6
jupyterlab_server~=2.27.3
notebook_shim~=0.2.4
et_xmlfile~=2.0.0
annotated-types~=0.7.0
semchunk~=2.2.2
tokenizers~=0.21.1
tenacity~=9.0.0
configobj~=5.0.9
asttokens~=3.0.0
executing~=2.2.0
ipykernel~=6.29.5
jsonlines~=3.1.0
langchain-core~=0.3.48
langsmith~=0.3.18
langchain-text-splitters~=0.3.7
requests-toolbelt~=1.0.0
nbconvert~=7.16.6
pandocfilters~=1.5.1
mistune~=3.1.3
jupyterlab_pygments~=0.3.0
pure_eval~=0.2.3
pycparser~=2.22
RapidFuzz~=3.12.2
websocket-client~=1.8.0
terminado~=0.18.1
jsonref~=1.1.0
referencing~=0.36.2
webcolors~=24.11.1
jsonpointer~=3.0.0
uri-template~=1.3.0
isoduration~=20.11.0
rfc3339-validator~=0.1.4
jsonschema-specifications~=2024.10.1
async-lru~=2.0.5
langdetect~=1.0.9
markdownify~=1.1.0
Send2Trash~=1.8.3
requests-file~=2.1.0
humanfriendly~=10.0
coloredlogs~=15.0.1
jupyter-lsp~=2.2.5
marshmallow~=3.26.1
flatbuffers~=25.2.10
safetensors~=0.5.3
latex2mathml~=3.77.0
unstructured~=0.17.2
unstructured-client~=0.31.3
python-iso639~=2025.2.18
wrapt~=1.17.2
jupyter-events~=0.12.0
python-json-logger~=3.3.0
prometheus_client~=0.21.1
jupyter_server_terminals~=0.5.3
jsonpatch~=1.33
typing-inspect~=0.9.0
dataclasses-json~=0.6.7
tomli~=2.2.1
mypy-extensions~=1.0.0
eval_type_backport~=0.2.2
httpx-sse~=0.4.0
nest-asyncio~=1.6.0